[GUI] remove override Tab/TabPane
