+++ gui/src/main/java/org/astraea/gui/tab/SettingNode.java
@@ -25,15 +25,15 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Properties;
 import java.util.stream.Collectors;
+import javafx.scene.Node;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.AsyncAdmin;
 import org.astraea.gui.Context;
 import org.astraea.gui.pane.PaneBuilder;
-import org.astraea.gui.pane.Tab;
 import org.astraea.gui.text.EditableText;
 import org.astraea.gui.text.NoneditableText;
 
-public class SettingTab {
+public class SettingNode {
 
   private static final String BOOTSTRAP_SERVERS = "bootstrap servers";
   private static final String JMX_PORT = "jmx port";
@@ -81,58 +81,52 @@ public class SettingTab {
             });
   }
 
-  public static Tab of(Context context) {
+  public static Node of(Context context) {
     var bootstrapKey = "bootstrap";
     var jmxPortKey = "jmx";
     var properties = loadProperty();
-    var pane =
-        PaneBuilder.of()
-            .input(
-                NoneditableText.highlight(BOOTSTRAP_SERVERS),
-                EditableText.singleLine()
-                    .defaultValue(properties.get(bootstrapKey))
-                    .disallowEmpty()
-                    .build())
-            .input(
-                NoneditableText.of(JMX_PORT),
-                EditableText.singleLine()
-                    .onlyNumber()
-                    .defaultValue(properties.get(jmxPortKey))
-                    .build())
-            .clickName("CHECK")
-            .clickListener(
-                (input, logger) -> {
-                  var bootstrapServers = input.nonEmptyTexts().get(BOOTSTRAP_SERVERS);
-                  Objects.requireNonNull(bootstrapServers);
-                  var jmxPort =
-                      Optional.ofNullable(input.nonEmptyTexts().get(JMX_PORT))
-                          .map(Integer::parseInt);
-                  saveProperty(
-                      Map.of(
-                          bootstrapKey,
-                          bootstrapServers,
-                          jmxPortKey,
-                          jmxPort.map(String::valueOf).orElse("")));
-                  var newAdmin = AsyncAdmin.of(bootstrapServers);
-                  return newAdmin
-                      .nodeInfos()
-                      .thenAccept(
-                          nodeInfos -> {
-                            context.replace(newAdmin);
-                            if (jmxPort.isEmpty()) {
-                              logger.log("succeed to connect to " + bootstrapServers);
-                              return;
-                            }
-                            context.replace(nodeInfos, jmxPort.get());
-                            logger.log(
-                                "succeed to connect to "
-                                    + bootstrapServers
-                                    + ", and jmx: "
-                                    + jmxPort.get()
-                                    + " works well");
-                          });
-                })
-            .build();
-    return Tab.of("setting", pane);
+    return PaneBuilder.of()
+        .input(
+            NoneditableText.highlight(BOOTSTRAP_SERVERS),
+            EditableText.singleLine()
+                .defaultValue(properties.get(bootstrapKey))
+                .disallowEmpty()
+                .build())
+        .input(
+            NoneditableText.of(JMX_PORT),
+            EditableText.singleLine().onlyNumber().defaultValue(properties.get(jmxPortKey)).build())
+        .clickName("CHECK")
+        .clickListener(
+            (input, logger) -> {
+              var bootstrapServers = input.nonEmptyTexts().get(BOOTSTRAP_SERVERS);
+              Objects.requireNonNull(bootstrapServers);
+              var jmxPort =
+                  Optional.ofNullable(input.nonEmptyTexts().get(JMX_PORT)).map(Integer::parseInt);
+              saveProperty(
+                  Map.of(
+                      bootstrapKey,
+                      bootstrapServers,
+                      jmxPortKey,
+                      jmxPort.map(String::valueOf).orElse("")));
+              var newAdmin = AsyncAdmin.of(bootstrapServers);
+              return newAdmin
+                  .nodeInfos()
+                  .thenAccept(
+                      nodeInfos -> {
+                        context.replace(newAdmin);
+                        if (jmxPort.isEmpty()) {
+                          logger.log("succeed to connect to " + bootstrapServers);
+                          return;
+                        }
+                        context.replace(nodeInfos, jmxPort.get());
+                        logger.log(
+                            "succeed to connect to "
+                                + bootstrapServers
+                                + ", and jmx: "
+                                + jmxPort.get()
+                                + " works well");
+                      });
+            })
+        .build();
   }
 }
